copy-writer/requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langsmith>=0.0.80
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
textstat>=0.7.0
yake>=0.4.8
python-readability>=0.2.0
fastapi>=0.100.0
uvicorn>=0.20.0
colorama>=0.4.6